requirements.txt
Pillow
numpy
pandas
requests
streamlit
